contract: add init-faucet script
Use the script to initialize the spl-token-faucet (creating the mint
Gh9ZwEmdLJ8DscKNTkTqPbNwLNNBjuSzaG9Vp2KGtKJr) after deploying the
spl-token-faucet program with program ID
4sN8PnN2ki2W4TFXAfzR645FWs8nimmsYeNtxM8RBK6A on a local validator.

An alternative is to store a dump of the mint account and use the
--bpf-program and --account flags of solana-test-validator to include
both the spl-token-faucet program and mint at validator startup. I will
probably add better support for that method in the future.